Conflicts:
	check/check_cluster.sh

+++ src/scip/misc.c
@@ -8037,7 +8037,7 @@ int SCIPmemccpy(
 
       while( cnt-- && (*destination++ = *src++) != stop ); /*lint !e722*/
 
-      return (destination - dest);
+      return (int)(destination - dest);
    }
 }
